Code Refactoring and Bug fixes
