+++ lib/screens/publish_screen.dart
@@ -23,7 +23,6 @@ class PublishScreen extends StatefulWidget {
 }
 
 class _PublishScreenState extends State<PublishScreen> {
-
   late EaselProvider provider;
 
   @override
@@ -46,15 +45,14 @@ class _PublishScreenState extends State<PublishScreen> {
                   children: [
                     Stack(
                       children: [
-                        if (provider.nftFormat.format == kImageText) ...[
-                          ImageWidget(file: provider.file!)
-                        ],
+                        if (provider.nftFormat.format == kImageText) ...[ImageWidget(file: provider.file!)],
                         if (provider.nftFormat.format == kVideoText) ...[
-                          VideoWidget(file: provider.file!, previewFlag: true,)
+                          VideoWidget(
+                            file: provider.file!,
+                            previewFlag: true,
+                          )
                         ],
-                        if (provider.nftFormat.format == kAudioText) ...[
-                          AudioWidget(file: provider.file!)
-                        ],
+                        if (provider.nftFormat.format == kAudioText) ...[AudioWidget(file: provider.file!)],
                         Positioned(
                             top: 60,
                             right: 10,
@@ -76,11 +74,10 @@ class _PublishScreenState extends State<PublishScreen> {
                         children: [
                           Text(
                             provider.artNameController.text,
-                            style:
-                                Theme.of(context).textTheme.headline5!.copyWith(
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w600,
-                                    ),
+                            style: Theme.of(context).textTheme.headline5!.copyWith(
+                                  color: Colors.black,
+                                  fontWeight: FontWeight.w600,
+                                ),
                           ),
                           const VerticalSpace(
                             4,
@@ -88,18 +85,10 @@ class _PublishScreenState extends State<PublishScreen> {
                           RichText(
                             text: TextSpan(
                                 text: "$kCreatedByText ",
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .bodyText2!
-                                    .copyWith(
+                                style: Theme.of(context).textTheme.bodyText2!.copyWith(
                                       fontSize: 20,
                                     ),
-                                children: [
-                                  TextSpan(
-                                      text: provider.artistNameController.text,
-                                      style: const TextStyle(
-                                          color: EaselAppTheme.kBlue))
-                                ]),
+                                children: [TextSpan(text: provider.artistNameController.text, style: const TextStyle(color: EaselAppTheme.kBlue))]),
                           ),
                           const Divider(
                             height: 40,
@@ -107,71 +96,61 @@ class _PublishScreenState extends State<PublishScreen> {
                           ),
                           Text(
                             kNftDetailsText,
-                            style:
-                                Theme.of(context).textTheme.bodyText2!.copyWith(
-                                      fontSize: 18,
-                                    ),
+                            style: Theme.of(context).textTheme.bodyText2!.copyWith(
+                                  fontSize: 18,
+                                ),
                           ),
                           Text(
                             provider.descriptionController.text,
-                            style:
-                                Theme.of(context).textTheme.caption!.copyWith(
-                                      fontSize: 14,
-                                    ),
+                            style: Theme.of(context).textTheme.caption!.copyWith(
+                                  fontSize: 14,
+                                ),
                           ),
                           const VerticalSpace(
                             10,
                           ),
                           Text(
                             "$kPriceText: ${provider.priceController.text.trim()} ${provider.selectedDenom.name}",
-                            style:
-                                Theme.of(context).textTheme.caption!.copyWith(
-                                      fontSize: 14,
-                                    ),
+                            style: Theme.of(context).textTheme.caption!.copyWith(
+                                  fontSize: 14,
+                                ),
                           ),
                           if (provider.nftFormat.format != kAudioText) ...[
                             Text(
                               "$kSizeText: ${provider.fileWidth} x ${provider.fileHeight}px ${provider.fileExtension.toUpperCase()}",
-                              style:
-                                  Theme.of(context).textTheme.caption!.copyWith(
-                                        fontSize: 14,
-                                      ),
+                              style: Theme.of(context).textTheme.caption!.copyWith(
+                                    fontSize: 14,
+                                  ),
                             )
                           ],
-                          if (provider.nftFormat.format == kVideoText ||
-                              provider.nftFormat.format == kAudioText) ...[
+                          if (provider.nftFormat.format == kVideoText || provider.nftFormat.format == kAudioText) ...[
                             Text(
                               "$kDurationText: ${provider.fileDuration / kSecInMillis} sec",
-                              style:
-                                  Theme.of(context).textTheme.caption!.copyWith(
-                                        fontSize: 14,
-                                      ),
+                              style: Theme.of(context).textTheme.caption!.copyWith(
+                                    fontSize: 14,
+                                  ),
                             )
                           ],
                           Text(
                             "$kDateText: ${getDate()}",
-                            style:
-                                Theme.of(context).textTheme.caption!.copyWith(
-                                      fontSize: 14,
-                                    ),
+                            style: Theme.of(context).textTheme.caption!.copyWith(
+                                  fontSize: 14,
+                                ),
                           ),
                           const VerticalSpace(
                             10,
                           ),
                           Text(
                             "$kNoOfEditionText: ${provider.noOfEditionController.text}",
-                            style:
-                                Theme.of(context).textTheme.caption!.copyWith(
-                                      fontSize: 14,
-                                    ),
+                            style: Theme.of(context).textTheme.caption!.copyWith(
+                                  fontSize: 14,
+                                ),
                           ),
-
                           Text(
                             "$kRoyaltyText: ${provider.royaltyController.text}%",
-                            style:
-                                Theme.of(context).textTheme.caption!.copyWith(
-                                      fontSize: 14,
-                                    ),
+                            style: Theme.of(context).textTheme.caption!.copyWith(
+                                  fontSize: 14,
+                                ),
                           ),
                         ],
                       ),
