+++ lib/widgets/video_widget_full_screen.dart
@@ -4,6 +4,7 @@ import 'dart:ui';
 import 'package:easel_flutter/easel_provider.dart';
 import 'package:easel_flutter/utils/constants.dart';
 import 'package:easel_flutter/utils/easel_app_theme.dart';
+import 'package:easel_flutter/widgets/video_builder.dart';
 import 'package:easel_flutter/widgets/video_progress_widget.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:provider/provider.dart';
@@ -24,35 +25,28 @@ class VideoWidgetFullScreen extends StatelessWidget {
         value: easelProvider,
         child: Stack(
           children: [
-            easelProvider.isVideoLoading
-                ? const Center(
-                    child: CircularProgressIndicator(
-                      strokeWidth: 2,
-                      valueColor: AlwaysStoppedAnimation<Color>(EaselAppTheme.kWhite),
-                    ),
-                  )
-                : easelProvider.videoLoadingError.isNotEmpty
-                    ? Center(
-                        child: Padding(
-                        padding: EdgeInsets.all(10.h),
-                        child: Text(
-                          videoPlayerError,
-                          style: TextStyle(fontSize: 18.sp, color: EaselAppTheme.kWhite),
-                        ),
-                      ))
-                    : easelProvider.videoPlayerController.value.isInitialized
-                        ? Center(
-                            child: AspectRatio(
-                              aspectRatio: easelProvider.videoPlayerController.value.aspectRatio,
-                              child: VideoPlayer(easelProvider.videoPlayerController),
-                            ),
-                          )
-                        : const Center(
-                            child: CircularProgressIndicator(
-                              strokeWidth: 2,
-                              valueColor: AlwaysStoppedAnimation<Color>(EaselAppTheme.kWhite),
-                            ),
-                          ),
+            VideoBuilder(
+                onVideoLoading: (BuildContext context) => const Center(
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                    valueColor: AlwaysStoppedAnimation<Color>(EaselAppTheme.kBlack),
+                  ),
+                ),
+                onVideoHasError: (BuildContext context) => Center(
+                    child: Padding(
+                      padding: const EdgeInsets.all(10),
+                      child: Text(
+                        videoPlayerError,
+                        style: TextStyle(fontSize: 18.sp, color: EaselAppTheme.kBlack),
+                      ),
+                    )),
+                onVideoInitialized: (BuildContext context) => Center(
+                  child: AspectRatio(
+                    aspectRatio: easelProvider.videoPlayerController.value.aspectRatio,
+                    child: VideoPlayer(easelProvider.videoPlayerController),
+                  ),
+                ),
+                easelProvider: easelProvider),
             Column(
               mainAxisAlignment: MainAxisAlignment.end,
               children: [
